pyproject.toml
[project]
name = "gym-trainerbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "psycopg>=3.3.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "djangorestframework>=3.15.0",
    "drf-spectacular>=0.28.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=24.0",
    "django-unfold>=0.40.0",
    "google-auth>=2.29.0",
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "django-storages>=1.14.0",
    "boto3>=1.34.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff.lint.per-file-ignores]
"src/core/settings/__init__.py" = ["F403"]
"src/system/apis/__init__.py" = ["F403"]
